Removed standalone Makefile
Now that cmake support is in, prototype uiJuce code has been added, and
a c++ conversion might happen, the old Makefile will bitrot.  Also, I
mistakenly invoked 'make' in the src/ directory once, which is a more common
operation than running cmake.  With a missing Makefile that would not have
"worked".  So, getting rid of the old Makefile.  Thanks to version control,
We can of course bring it back if desired.