Refactor MoveList (remote MLIST_MAX_MOVES limit)
Makes MoveList its own class, and (to support future variants) removes the
hard limit on the number of moves.  move limit.  Instead, moves are inserted
into a std::vector.  The vectors are reused so we do not spend all of our
time dynamically allocating memory.

Unfortunately there is a roughly 7% slowdown from these changes.  AFAICT
this is due to cache effects -- the moves are added to heap memory instead
of the stack, for starters.  (Repeats mantra: Make it work, then make it
fast ...)

Other moderate cleanup happened as well.