Refactor GameT into a 'Game' C++ class (+ many other changes)
This began as an effort to move the HintPv from gVars to somewhere under
the engine code, then exploded.  It became apparent I would need to
refactor the Game struct into a class so that it could only be manipulated
in a controlled fashion.  This touched a lot of code and led to discovery
of several other issues.

Other issues fixed (+ probably some that I forgot):
uiXboard: fix poorly formatted error strings (no newline)
fix think time w/infinite clock but per-move limit.
fixed crash when switching on the fly from xboard to uci interface.
uiUci: better attempts to record a bestMove when we stop pondering; handle
    'ponder + searchmoves' sanely; more correct ponder info and score
fixed HintPv::Rewind() to not clobber memory (I had a bad understanding of
    std::copy_backward()).