Various unrelated bugfixes
Fix a recently introduced crash when telling the searchers to stop thinking
(due to an improperly thought-out assert()). This was fairly serious.
Fix SAN move disambiguation.  A previous related attempt in version 1.0 was
too quick to add an unnecessary file and/or rank qualifier.
Fix a silly "Error: Error: (...)" uiNcurses print when we attempt to exit
the Edit menu with an illegal position.