Reorganize/split Thinker into (Engine, Thinker)
This is an initial effort to split the Thinker's functionality into the
subset that should be exposed to upper layers (the 'Engine') and the subset
that should run in its own thread (the Thinker proper).

To be fair, the Engine's implementation currently still does an awful lot of
reaching into the Thinker class to do what it needs to.