Refactored ListT into C++ class
It is now known as arctic::List (and arctic::ListElement).

This came with a ~1.2% speedup, mostly due to the fact that we now disable
list sanity checking by default.  Obviously that makes things much harder
to debug if we have list corruption.  The checks can be re-enabled by
is encountered.