Refactored MoveList object cache to be a generic template
Now, anyone can have an object cache of any type they want.  The cache is
not threadsafe, but could be made so if necessary.

I removed the 'exiting' functionality, since we don't need it anymore and
it appeared (from http://en.cppreference.com/w/cpp/language/lifetime,
specifically "Access outside of lifetime") that I was relying on undefined
behavior.