Implement EventQueue, Pollable, SimpleQueue, Timer
This is first in (hopefully) a series of steps to move away from unix
sockets for communicating messages between Thinker threads, and to use an
EventQueue instead.

The motivations are:
-- I can pass complex objects through an EventQueue w/out serialization
-- Write atomicity from different threads is never a concern.  This in turn
   provides a cleaner way to move the Thinking timer directly under the
   control of the Thinker (timer thread pops and enqueues a timeout event).
-- For any eventual windows port, I can avoid having to either abstract
   over an anonymous pipe, or using full-blown TCP loopback socket.

For whatever reason (memory layout?  Suboptimal performance?) this is
slightly (.05%?) slower than the socket route so far -- and it will
probably get worse -- but so far, given the potential for clean code, I
think it is worth the hit.

I am a little worried about potential heap allocation of the EventQueue
elements, though.

Class descriptions:

EventQueue implements a boost::asio::io_service-like object that one can
post to.  The advantage of an EventQueue is a quicker emptiness check (no
mutex is taken), and better composibility thanks to Pollable.

SimpleQueue can be given to std::queue, and incorporates a queue discipline
on top of a std::vector-like array.

Timer implements a timer, whose expiration handler runs on a dedicated
thread.

As an unrelated change, I sped up arctic::Semaphore slightly by making
post()ers release the internal mutex before signaling the waiter
(otherwise, the waiter may wake up only to wait again (briefly) on that
mutex).