Make Thinker able to move itself due to time limits
Now, when the Engine is told to think, it can move itself after calculating
how much time it wants to use.  We use an internal Timer instead of the
playloop's polling loop to achieve this.

In the future, this should make it easier to adjust our thinking time based
on the situation (for example, allot more time if we have a last-second
fail-low), and is also another (small) step towards being able to interface
with external engines.

The code introduces a negligible (but noticeable) slowdown.