Minor cleanup (Make ui func table more object-y)
The idea is that eventually the UI func table will become a view of sorts.  We
do assume that the view is only interested in one game, but since switching
the game out from under the view is almost certainly not handled well, I think
it's a good cleanup for now.

Oddly enough this appears to cause a minor slowdown, probably due to caching
effects.  Too bad.